+++ modules/asset/group/src/Form/AssetGroupActionForm.php
@@ -138,7 +138,7 @@ class AssetGroupActionForm extends ConfirmFormBase {
     $form['date'] = [
       '#type' => 'datetime',
       '#title' => $this->t('Date'),
-      '#default_value' => new DrupalDateTime('midnight'),
+      '#default_value' => new DrupalDateTime('midnight', $this->user->getTimeZone()),
       '#required' => TRUE,
     ];