+++ modules/quick/planting/src/Plugin/QuickForm/Planting.php
@@ -11,6 +11,7 @@ use Drupal\Core\Session\AccountInterface;
 use Drupal\Core\State\StateInterface;
 use Drupal\farm_quick\Plugin\QuickForm\QuickFormBase;
 use Drupal\farm_quick\Traits\QuickAssetTrait;
+use Drupal\farm_quick\Traits\QuickFormElementsTrait;
 use Drupal\farm_quick\Traits\QuickLogTrait;
 use Drupal\farm_quick\Traits\QuickQuantityTrait;
 use Drupal\farm_quick\Traits\QuickStringTrait;
@@ -38,6 +39,7 @@ class Planting extends QuickFormBase {
   use QuickLogTrait;
   use QuickQuantityTrait;
   use QuickStringTrait;
+  use QuickFormElementsTrait;
 
   /**
    * The entity type manager service.
@@ -328,64 +330,61 @@ class Planting extends QuickFormBase {
     }
 
     // Create log fields.
-    $field_info = [
-      'date' => [
-        '#type' => 'datetime',
-        '#title' => $this->t('Date'),
-        '#default_value' => new DrupalDateTime('midnight', $this->currentUser->getTimeZone()),
-        '#required' => TRUE,
-      ],
-      'done' => [
-        '#type' => 'checkbox',
-        '#title' => $this->t('Completed'),
-      ],
-      'location' => [
-        '#type' => 'entity_autocomplete',
-        '#title' => $this->t('Location'),
-        '#description' => $this->t('Where does this take place?'),
-        '#target_type' => 'asset',
-        '#selection_handler' => 'views',
-        '#selection_settings' => [
-          'view' => [
-            'view_name' => 'farm_location_reference',
-            'display_name' => 'entity_reference',
-            'arguments' => [],
-          ],
-          'match_operator' => 'CONTAINS',
+    $field_info = [];
+    $field_info['date'] = [
+      '#type' => 'datetime',
+      '#title' => $this->t('Date'),
+      '#default_value' => new DrupalDateTime('midnight', $this->currentUser->getTimeZone()),
+      '#required' => TRUE,
+    ];
+    $field_info['done'] = [
+      '#type' => 'checkbox',
+      '#title' => $this->t('Completed'),
+    ];
+    $field_info['location'] = [
+      '#type' => 'entity_autocomplete',
+      '#title' => $this->t('Location'),
+      '#description' => $this->t('Where does this take place?'),
+      '#target_type' => 'asset',
+      '#selection_handler' => 'views',
+      '#selection_settings' => [
+        'view' => [
+          'view_name' => 'farm_location_reference',
+          'display_name' => 'entity_reference',
+          'arguments' => [],
         ],
-        '#required' => TRUE,
+        'match_operator' => 'CONTAINS',
       ],
-      'quantity' => [
-        '#type' => 'details',
-        '#title' => $this->t('Quantity'),
-        '#open' => FALSE,
-        'measure' => [
-          '#type' => 'select',
-          '#title' => $this->t('Measure'),
-          '#options' => $filtered_quantity_measure_options,
-          '#default_value' => 'weight',
-        ],
-        'value' => [
-          '#type' => 'textfield',
-          '#title' => $this->t('Value'),
-        ],
-        'units' => [
-          '#type' => 'entity_autocomplete',
-          '#title' => $this->t('Units'),
-          '#target_type' => 'taxonomy_term',
-          '#selection_settings' => [
-            'target_bundles' => ['unit'],
-          ],
-          '#autocreate' => [
-            'bundle' => 'unit',
-          ],
-        ],
+      '#required' => TRUE,
+    ];
+    $field_info['quantity'] = $this->buildInlineContainer();
+    $field_info['quantity']['value'] = [
+      '#type' => 'textfield',
+      '#title' => $this->t('Quantity'),
+      '#size' => 16,
+    ];
+    $field_info['quantity']['units'] = [
+      '#type' => 'entity_autocomplete',
+      '#title' => $this->t('Units'),
+      '#target_type' => 'taxonomy_term',
+      '#selection_settings' => [
+        'target_bundles' => ['unit'],
       ],
-      'notes' => [
-        '#type' => 'text_format',
-        '#title' => $this->t('Notes'),
-        '#format' => 'default',
+      '#autocreate' => [
+        'bundle' => 'unit',
       ],
+      '#size' => 16,
+    ];
+    $field_info['quantity']['measure'] = [
+      '#type' => 'select',
+      '#title' => $this->t('Measure'),
+      '#options' => $filtered_quantity_measure_options,
+      '#default_value' => 'weight',
+    ];
+    $field_info['notes'] = [
+      '#type' => 'text_format',
+      '#title' => $this->t('Notes'),
+      '#format' => 'default',
     ];
     foreach ($include_fields as $field) {
       if (array_key_exists($field, $field_info)) {