+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [Fix quick form toolbar menu links #657](https://github.com/farmOS/farmOS/pull/657)
+- [Respect user timezone in midnight default datetime values #665](https://github.com/farmOS/farmOS/pull/665)
 
 ## [2.0.3] 2023-03-15